requirements.txt
streamlit>=1.25.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-google-genai>=2.0.0
langchain-openai>=0.3.0
python-docx>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
python-dotenv>=1.0.0
